2024.01.02 12:21pm css 수정

+++ farmfarm/src/main/java/com/farmfarm/dto/MyPageUserCartVO.java
@@ -26,6 +26,6 @@ public class MyPageUserCartVO {
 	int min_amount;
 	int d_day;
 	int current_price;
-	int total_pct;
+	double total_pct;
 	int participants;
 }
